file_current_offset is now a long and fix scrolling

+++ include/draw.h
@@ -10,12 +10,12 @@
 
 void gui_init();
 void gui_draw_title(char *format, ...);
-void gui_draw_hex(byte *file, unsigned int file_current_offset);
-void draw_cursor_reset(unsigned int *file_current_offset);
-void draw_cursor_up(unsigned int *file_current_offset, byte *file);
-void draw_cursor_down(unsigned int *file_current_offset, byte *file);
-void draw_cursor_right(unsigned int *file_current_offset, byte *file);
-void draw_cursor_left(unsigned int *file_current_offset, byte *file);
+void gui_draw_hex(byte *file, long file_current_offset);
+void draw_cursor_reset(long *file_current_offset);
+void draw_cursor_up(long *file_current_offset, byte *file);
+void draw_cursor_down(long *file_current_offset, byte *file);
+void draw_cursor_right(long *file_current_offset, byte *file);
+void draw_cursor_left(long *file_current_offset, byte *file);
 void gui_end();
 
 #endif
